(Feat) Overhaul stats code
